rework module file structure
