+++ contrib/impcap/icmp_parser.c
@@ -1,0 +1,16 @@
+#include "parser.h"
+
+void handle_icmp_header(const uchar *packet, size_t pktSize, struct json_object *jparent) {
+  DBGPRINTF("handle_icmp_header\n");
+  DBGPRINTF("packet size %d\n", pktSize);
+
+  if(pktSize < 8) {
+    DBGPRINTF("ICMP packet too small : %d\n", pktSize);
+    return;
+  }
+
+  icmp_header_t *icmp_header = (icmp_header_t *)packet;
+
+  json_object_object_add(jparent, "ICMP_type", json_object_new_int(icmp_header->type));
+  json_object_object_add(jparent, "ICMP_code", json_object_new_int(icmp_header->code));
+}
